# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Экскурсии по Удмуртской Республике",
        "description": "API документация для работы с сервисом экскурсий по УР",
        "version": "0.0.1"
    },
    "paths": {
        "/api/poi/create": {
            "post": {
                "tags": [
                    "Достопримечательности"
                ],
                "summary": "Создать новую достопримечательность",
                "description": "Создаёт и возвращает новую достопримечательность в случае успеха",
                "operationId": "b3a9db44be281c210db1dc5686205879",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "description",
                                    "file"
                                ],
                                "properties": {
                                    "file": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Описание достопримечательности"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Успешное создание",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "POI created"
                                        },
                                        "data": {
                                            "properties": {
                                                "description": {
                                                    "type": "string",
                                                    "example": "Описание достопримечательности"
                                                },
                                                "file_name": {
                                                    "type": "string",
                                                    "example": "pois/wydMNczrMWlmuNRPbB2jmO5PBzOEsPkm7mfchVng.png"
                                                },
                                                "user_id": {
                                                    "type": "number",
                                                    "example": 1
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "foo@bar.com"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2000-01-01T00:00:00.000000Z"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2000-01-01T00:00:00.000000Z"
                                                },
                                                "id": {
                                                    "type": "number",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации данных",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation error"
                                        },
                                        "errors": {
                                            "properties": {
                                                "description": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "example": [
                                                        "The description field is required."
                                                    ]
                                                },
                                                "file": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "example": [
                                                        "The file field is required."
                                                    ]
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Пользователь не авторизирован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/poi/list": {
            "get": {
                "tags": [
                    "Достопримечательности"
                ],
                "summary": "Получить список достопримечательностей",
                "description": "Возвращает список всех достопримечательностей",
                "operationId": "7379957e55709238e32fe7d749b26e30",
                "responses": {
                    "200": {
                        "description": "Успешный вывод достопримечательностей",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "POIs retrieved"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Описание достопримечательности"
                                                    },
                                                    "file_name": {
                                                        "type": "string",
                                                        "example": "pois/wydMNczrMWlmuNRPbB2jmO5PBzOEsPkm7mfchVng.png"
                                                    },
                                                    "user_id": {
                                                        "type": "number",
                                                        "example": 1
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "foo@bar.com"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2000-01-01T00:00:00.000000Z"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2000-01-01T00:00:00.000000Z"
                                                    },
                                                    "id": {
                                                        "type": "number",
                                                        "example": 1
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Пользователь не авторизирован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/content/udmurtia": {
            "get": {
                "tags": [
                    "Контент"
                ],
                "summary": "Контент по Удмуртии",
                "description": "Возвращает статический контент в формате markdown для страницы",
                "operationId": "910572190b5a8e97c01de37e7f49653d",
                "responses": {
                    "200": {
                        "description": "Успешный вывод контента",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Content fetched"
                                        },
                                        "data": {
                                            "properties": {
                                                "name": {
                                                    "type": "string",
                                                    "example": "about-udmurtia"
                                                },
                                                "content": {
                                                    "type": "string",
                                                    "example": "Lorem ipsum"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2000-01-01T00:00:00.000000Z"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2000-01-01T00:00:00.000000Z"
                                                },
                                                "id": {
                                                    "type": "number",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Пользователь не авторизирован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}